add work logs controller and import method

+++ db/migrate/20180811150903_create_work_logs.rb
@@ -2,7 +2,7 @@ class CreateWorkLogs < ActiveRecord::Migration[5.2]
   def change
     create_table :work_logs do |t|
       t.date :date
-      t.decimal :hours_worked
+      t.float :hours_worked
       t.integer :upload_number
       t.references :employee, foreign_key: true
